A secondary purpose within the app, even if it's a json list of Top most controversial or top 5 latest votes. { cp.avatarUrl }- { cp.party } | {cp.name }  { congressperson.state }
How do we trigger two separate HTTP events simultaneously, and let's be real, It would be ideal to call your senator right this second, but i'm going to call in 300 seconds or 120 secs or 30 secs. The less time the less room for distraction and waste of resources. maybe using what I think is called WebHelper to essentially hook and prioritize calls, and finishing with a slow post to the database and ip, cookie, user, phone number or whatever info that could keep from having to fetch all politicians from their state and zip. What's everyone think about the votes thing. It's something that doesnt have to be THIS, it could be an add-on like it should, with a register2Vote API Link + social links @following groups interest perhaps or maybe not. check @DevProgress